backend/requirements.txt
fastapi
uvicorn

python-jose[cryptography]

passlib[bcrypt]
bcrypt<4.1.0

sqlalchemy
databases[postgresql]
psycopg2-binary

pydantic
email-validator

python-multipart
